+++ lib/href_page_title.rb
@@ -8,17 +8,17 @@ module MSlinn
       all_pages = @site.all_collections + @site.pages
       pages = if @match
                 all_pages.select { |page| page.url&.include? path }
-              elsif @label_source == :from_page_title
+              elsif @label_source == LabelSource::FROM_PAGE_TITLE
                 all_pages.select { |page| page.url == path }
               end
       pages&.first
     end
 
+    # Handles links to Jekyll pages
+    # Uses the linked page title as the link text
     def handle_page_title(linkk)
       @follow = @target = ''
-      @external_link = linkk.start_with? 'http'
-      @local_link = !@external_link
-      raise HRefError, 'href tags with page_title require local links.' unless @local_link
+      raise HRefError, 'href tags with page_title require local links.' unless @link_type == LinkType::LOCAL
 
       page = find_page linkk
       unless page
@@ -27,22 +27,7 @@ module MSlinn
         raise HRefError, msg
       end
       @text = @label = page.title
-    rescue HRefError => e
-      msg = format_error_message "#{e.message}\n<pre>  {% href #{@argument_string.strip} %}</pre>"
-      @text = "<div class='href_error'>#{msg}</div>"
-      @link = linkk
-
-      e.shorten_backtrace
-      @logger.error { "#{e.class} raised #{JekyllPluginHelper.remove_html_tags msg}" }
-      raise e if @die_on_href_error
-
-      "<div class='href_error'>HRefError raised in #{self.class};\n#{msg}</div>"
-    ensure
-      if @text.to_s.empty?
-        handle_empty_text linkk
-      else
-        handle_text linkk
-      end
+      handle_empty_text linkk if @text.to_s.empty?
       @label
     end
   end