+++ lib/href_summary.rb
@@ -6,9 +6,9 @@ module MSlinn
       return if @summary_exclude || @link_save.start_with?('mailto:') || @link_save.start_with?('#')
 
       @summary = @summary.to_s.empty? ? @text : @summary.to_s
-      if @summary == true
+      if @summary.to_s == 'true'
         warning = <<~END_WARNING
-          Warning: a href plugin keyword option was detected in the link text for #{@path} on line #{line_number}.
+          Warning: The 'summary' href plugin keyword option was detected in the link text for #{@path} on line #{line_number}.
           The href tag will not be included in the summary, and the link text will not have the word summary included.
           This is probably unintentional. Consider using the label option to correct this problem."
         END_WARNING