+++ lib/jekyll_href.rb
@@ -1,7 +1,7 @@
 require_relative 'href_tag'
 require_relative 'href_summary_tag'
 
-HrefError = Class.new(Liquid::Error)
+HrefError = Class.new Liquid::Error
 
 module JekyllHrefModule
   include MSlinn