+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Change Log
 
+## 1.2.8 / 2023-12-31
+
+* Fixed missing ' rel="nofollow"' for links without labels.
+
 
 ## 1.2.7 / 2023-12-09